Add the possibility to have consistent uid for aptly user
If someone needs to recreate aptly over and over again, he would
probably want to have full consistency and control over the uid
used by aptly.

This patch adds the possibility to define aptly_uid. If defined,
the uid of aptly user will be set, if not the current behavior is
unchanged.