Ensure consistency of the become
In some tasks we have become and become_user, while in other there
is su - <user> -c. This could just be replaced by become,
become_user and become_method.

On top of that, the shell isn't necessary anymore, and command can
be used instead. That way it will not be the cause of ansible-lint
issues.